backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg[binary]==3.1.16
alembic==1.13.1

# Vector Store
qdrant-client==1.7.3

# OpenAI & AI
openai==1.10.0
tiktoken==0.5.2
langchain==0.1.0
langchain-openai==0.0.5

# Translation (Free alternative to Google Cloud)
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
# google-cloud-translate==3.15.0  # Optional paid service

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP & Utils
httpx==0.26.0
redis==5.0.1

# Markdown & Text Processing
python-frontmatter==1.1.0
markdown==3.5.2

# Voice & Audio (Free/Open-Source)
openai-whisper==20240314
edge-tts==6.1.6
pydub==0.25.1
librosa==0.10.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.1
